refactor(conjugation): rewrite relaxation workflow
