+++ src/polyzymd/builders/conjugation/pablo/__init__.py
@@ -29,7 +29,6 @@ from polyzymd.builders.conjugation.pablo.parameterization import (
     DEFAULT_CONJUGATION_FORCE_FIELD_NAMES,
     InterchangeParameterizationResult,
     InterchangeParameterizationSettings,
-    build_formal_charge_smoke_template,
     create_interchange_from_openff_topology,
     create_interchange_from_pablo_topology,
     deduplicate_charge_templates,
@@ -73,7 +72,6 @@ __all__ = [
     "ConjugationChargeConfig",
     "LocalChargePatchError",
     "ResiduePartialChargeRecord",
-    "build_formal_charge_smoke_template",
     "build_conjugate_charge_templates",
     "build_local_product_charge_patch_records",
     "build_pablo_residue_library",
