+++ src/polyzymd/builders/conjugation/relaxation/__init__.py
@@ -1,0 +1,20 @@
+"""Conjugate relaxation API."""
+
+from polyzymd.builders.conjugation.relaxation.models import (
+    ConjugateRelaxationDiagnostics,
+    ConjugateRelaxationResult,
+    ConjugateRelaxationSettings,
+    OpenMMValidationResult,
+    OpenMMValidationSettings,
+)
+from polyzymd.builders.conjugation.relaxation.openmm import relax_conjugate, validate_openmm_product
+
+__all__ = [
+    "ConjugateRelaxationDiagnostics",
+    "ConjugateRelaxationResult",
+    "ConjugateRelaxationSettings",
+    "OpenMMValidationResult",
+    "OpenMMValidationSettings",
+    "relax_conjugate",
+    "validate_openmm_product",
+]
