+++ tests/test_conjugation_gpu_solvation.py
@@ -62,7 +62,7 @@ def test_opt_in_cuda_required_for_conjugation_solvation_full_build(tmp_path: Pat
 
     This expensive test is intentionally opt-in and must not fall back to CPU.
     The linked conjugate input is generated by default and must complete the
-    direct link, restrained vacuum minimization, and short vacuum MD before the
+    direct link and short OpenMM product validation before the
     relaxed conformer is solvated through the OpenFF/PACKMOL/Interchange path.
     """
     openmm, openmm_unit, platform = _require_gpu_solvation_stack_or_skip()
@@ -92,32 +92,32 @@ def test_opt_in_cuda_required_for_conjugation_solvation_full_build(tmp_path: Pat
 
     from openff.interchange.interop.openmm._positions import to_openmm_positions
 
-    from polyzymd.builders.conjugation._relaxation import (
-        VacuumSmokeSettings,
-        run_restrained_vacuum_smoke,
-    )
     from polyzymd.builders.conjugation.pablo.parameterization import (
         create_interchange_from_openff_topology,
         set_topology_positions_from_pdb,
     )
+    from polyzymd.builders.conjugation.relaxation.openmm import (
+        OpenMMValidationSettings,
+        validate_openmm_product,
+    )
     from polyzymd.builders.solvent import SolventBuilder, SolventComposition
     from polyzymd.data.solvent_molecules import get_solvent_molecule
 
-    with _timed_stage("vacuum parameterization"):
-        vacuum_parameterization = create_interchange_from_openff_topology(
+    with _timed_stage("OpenMM validation parameterization"):
+        validation_parameterization = create_interchange_from_openff_topology(
             direct_result.topology,
             charge_from_molecules=(direct_result.charge_template,),
-            success_diagnostic="Direct linked conjugate was parameterized before vacuum relaxation",
+            success_diagnostic="Direct linked conjugate was parameterized before validation",
             failure_subject="direct linked conjugate topology",
         )
-    assert vacuum_parameterization.interchange is not None
+    assert validation_parameterization.interchange is not None
 
-    with _timed_stage("restrained vacuum EM and MD"):
-        vacuum_result = run_restrained_vacuum_smoke(
-            vacuum_parameterization.interchange,
-            artifact_dir / "vacuum-relaxation",
+    with _timed_stage("short OpenMM product validation"):
+        validation_result = validate_openmm_product(
+            validation_parameterization.interchange,
+            artifact_dir / "openmm-validation",
             protein_heavy_atom_indices=None,
-            settings=VacuumSmokeSettings(
+            settings=OpenMMValidationSettings(
                 minimization_max_iterations=_int_env(
                     GPU_MINIMIZATION_ITERS_ENV_VAR,
                     default=DEFAULT_GPU_MINIMIZATION_ITERS,
@@ -126,16 +126,15 @@ def test_opt_in_cuda_required_for_conjugation_solvation_full_build(tmp_path: Pat
                 platform_name="CUDA",
             ),
         )
-    assert vacuum_result.success
-    assert vacuum_result.equilibrated_pdb_path is not None
-    assert vacuum_result.equilibrated_pdb_path.exists()
-    assert vacuum_result.nvt_steps > 0, "Restrained vacuum MD must run before solvation"
-    assert vacuum_result.restrained_atom_count == expected_restrained_atom_count
+    assert validation_result.success
+    assert validation_result.validation_json_path.exists()
+    assert validation_result.nvt_steps > 0, "Short OpenMM validation must run before solvation"
+    assert validation_result.restrained_atom_count == expected_restrained_atom_count
 
     with _timed_stage("relaxed position transfer"):
         relaxed_topology = set_topology_positions_from_pdb(
             direct_result.topology,
-            vacuum_result.equilibrated_pdb_path,
+            direct_result.linked_pdb_path,
         )
 
     with _timed_stage("solvation box and count setup"):
@@ -219,7 +218,7 @@ def test_gpu_solvation_smoke_source_avoids_openmm_modeller_solvation():
         assert token not in source
 
 
-def test_gpu_solvation_smoke_source_uses_all_heavy_vacuum_restraints():
+def test_gpu_solvation_smoke_source_uses_all_heavy_validation_restraints():
     """The GPU path should not pass a protein-only restraint selection."""
     source = Path(__file__).read_text(encoding="utf-8")
     protein_only_call = "protein_heavy_atom_indices=" + "_protein_heavy_atom_indices"
@@ -229,7 +228,7 @@ def test_gpu_solvation_smoke_source_uses_all_heavy_vacuum_restraints():
 
 
 def test_gpu_solvation_nvt_steps_env_rejects_zero(monkeypatch: pytest.MonkeyPatch):
-    """The GPU solvation smoke must run short vacuum MD before solvation."""
+    """The GPU solvation smoke must run short OpenMM validation before solvation."""
     monkeypatch.setenv(GPU_NVT_STEPS_ENV_VAR, "0")
 
     with pytest.raises(ValueError, match=f"{GPU_NVT_STEPS_ENV_VAR} must be positive"):
@@ -344,7 +343,7 @@ def _direct_artifact_from_env_or_generate(artifact_dir: Path):
         pytest.fail(
             "POLYZYMD_CONJUGATION_LINKED_PDB is no longer accepted for the GPU solvation smoke. "
             "The test must generate a direct artifact with an OpenFF topology and charge template "
-            "so it can run vacuum relaxation before solvation."
+            "so it can run OpenMM product validation before solvation."
         )
 
     from tests.test_conjugation_integrated_direct_smoke import build_direct_smoke_artifact
