+++ tests/test_conjugation_validation.py
@@ -13,7 +13,7 @@ from polyzymd.builders.conjugation.validation import (
     ValidationStatus,
     audit_charge_reports,
     audit_linkage_geometry,
-    audit_openmm_smoke_reports,
+    audit_openmm_validation_reports,
     audit_parameter_coverage,
     build_conjugate_validation_report,
     validate_atom_presence,
@@ -588,31 +588,31 @@ def test_linkage_geometry_warns_for_nonbonded_close_contacts():
     assert any(check.name == "nonbonded_close_contacts" for check in report.checks)
 
 
-def test_smoke_audit_pass_fail_and_skipped(tmp_path):
-    """Smoke audit should consume fake JSON payloads."""
-    assert audit_openmm_smoke_reports(tmp_path).status == ValidationStatus.SKIPPED
+def test_openmm_validation_audit_pass_fail_and_skipped(tmp_path):
+    """OpenMM validation audit should consume fake JSON payloads."""
+    assert audit_openmm_validation_reports(tmp_path).status == ValidationStatus.SKIPPED
 
-    smoke_path = tmp_path / "vacuum_smoke.json"
-    smoke_path.write_text(
+    validation_path = tmp_path / "openmm_validation.json"
+    validation_path.write_text(
         json.dumps({"success": True, "energy_after_min_kj_mol": -1.0}),
         encoding="utf-8",
     )
-    assert audit_openmm_smoke_reports(tmp_path).status == ValidationStatus.PASS
+    assert audit_openmm_validation_reports(tmp_path).status == ValidationStatus.PASS
 
-    smoke_path.write_text(
+    validation_path.write_text(
         json.dumps({"success": True, "energy_after_min_kj_mol": "nan"}),
         encoding="utf-8",
     )
-    assert audit_openmm_smoke_reports(tmp_path).status == ValidationStatus.FAIL
+    assert audit_openmm_validation_reports(tmp_path).status == ValidationStatus.FAIL
 
 
-def test_smoke_audit_does_not_mask_failed_legacy_smoke_with_frozen_evidence(tmp_path):
-    """Frozen relaxation evidence must not suppress failed legacy smoke artifacts."""
-    (tmp_path / "restrained_smoke_diagnostics.json").write_text(
+def test_openmm_validation_audit_ignores_legacy_validation_artifacts(tmp_path):
+    """Legacy validation artifacts should not affect the new relaxation audit."""
+    (tmp_path / "legacy_openmm_validation.json").write_text(
         json.dumps({"success": False}),
         encoding="utf-8",
     )
-    diagnostics_path = tmp_path / "frozen_protein_relaxation_diagnostics.json"
+    diagnostics_path = tmp_path / "conjugate_relaxation.json"
     diagnostics_path.write_text(
         json.dumps(
             {
@@ -637,49 +637,48 @@ def test_smoke_audit_does_not_mask_failed_legacy_smoke_with_frozen_evidence(tmp_
         encoding="utf-8",
     )
 
-    report = audit_openmm_smoke_reports(tmp_path)
-
-    assert report.status == ValidationStatus.FAIL
-    assert report.frozen_relaxation_diagnostics_json_path == diagnostics_path
-    assert any(check.name == "restrained_smoke_diagnostics" for check in report.checks)
-
-
-def test_smoke_audit_accepts_passing_frozen_relaxation_evidence(tmp_path):
-    """Validation should pass when all present OpenMM smoke evidence passes."""
-    diagnostics_path = tmp_path / "frozen_protein_relaxation_diagnostics.json"
-    diagnostics_path.write_text(
-        json.dumps(
-            {
-                "success": True,
-                "stage_a_success": True,
-                "stage_b_success": True,
-                "temporary_anchor_count": 2,
-                "barostat_used": False,
-                "stage_a_energy_after_min_kj_mol": -10.0,
-                "stage_b_energy_after_md_kj_mol": -11.0,
-                "stage_b_protein_rmsd_from_stage_a_angstrom": 0.0,
-                "stage_b_protein_max_displacement_from_stage_a_angstrom": 0.0,
-                "stage_b_linkage_distance_errors_angstrom": [0.1, 0.2],
-                "settings": {
-                    "md_steps": 10,
-                    "max_protein_rmsd_angstrom": 0.05,
-                    "max_protein_displacement_angstrom": 0.25,
-                    "max_linkage_distance_error_angstrom": 0.35,
-                },
-            }
-        ),
-        encoding="utf-8",
-    )
-
-    report = audit_openmm_smoke_reports(tmp_path)
+    report = audit_openmm_validation_reports(tmp_path)
 
     assert report.status == ValidationStatus.PASS
-    assert report.frozen_relaxation_diagnostics_json_path == diagnostics_path
+    assert report.relaxation_diagnostics_json_path == diagnostics_path
 
 
-def test_smoke_audit_rejects_stale_zero_step_frozen_relaxation(tmp_path):
+def test_openmm_validation_audit_accepts_passing_conjugate_relaxation_evidence(tmp_path):
+    """Validation should pass when all present OpenMM validation evidence passes."""
+    diagnostics_path = tmp_path / "conjugate_relaxation.json"
+    diagnostics_path.write_text(
+        json.dumps(
+            {
+                "success": True,
+                "stage_a_success": True,
+                "stage_b_success": True,
+                "temporary_anchor_count": 2,
+                "barostat_used": False,
+                "stage_a_energy_after_min_kj_mol": -10.0,
+                "stage_b_energy_after_md_kj_mol": -11.0,
+                "stage_b_protein_rmsd_from_stage_a_angstrom": 0.0,
+                "stage_b_protein_max_displacement_from_stage_a_angstrom": 0.0,
+                "stage_b_linkage_distance_errors_angstrom": [0.1, 0.2],
+                "settings": {
+                    "md_steps": 10,
+                    "max_protein_rmsd_angstrom": 0.05,
+                    "max_protein_displacement_angstrom": 0.25,
+                    "max_linkage_distance_error_angstrom": 0.35,
+                },
+            }
+        ),
+        encoding="utf-8",
+    )
+
+    report = audit_openmm_validation_reports(tmp_path)
+
+    assert report.status == ValidationStatus.PASS
+    assert report.relaxation_diagnostics_json_path == diagnostics_path
+
+
+def test_openmm_validation_audit_rejects_stale_zero_step_conjugate_relaxation(tmp_path):
     """Validation should reject stale frozen diagnostics with no Stage B MD."""
-    (tmp_path / "frozen_protein_relaxation_diagnostics.json").write_text(
+    (tmp_path / "conjugate_relaxation.json").write_text(
         json.dumps(
             {
                 "success": True,
@@ -697,15 +696,15 @@ def test_smoke_audit_rejects_stale_zero_step_frozen_relaxation(tmp_path):
         encoding="utf-8",
     )
 
-    report = audit_openmm_smoke_reports(tmp_path)
+    report = audit_openmm_validation_reports(tmp_path)
 
     assert report.status == ValidationStatus.FAIL
-    assert any(check.name == "frozen_protein_relaxation_md_steps" for check in report.checks)
+    assert any(check.name == "conjugate_relaxation_md_steps" for check in report.checks)
 
 
-def test_smoke_audit_requires_frozen_relaxation_md_steps(tmp_path):
+def test_openmm_validation_audit_requires_conjugate_relaxation_md_steps(tmp_path):
     """Validation should reject otherwise-passing frozen diagnostics without MD steps."""
-    (tmp_path / "frozen_protein_relaxation_diagnostics.json").write_text(
+    (tmp_path / "conjugate_relaxation.json").write_text(
         json.dumps(
             {
                 "success": True,
@@ -727,15 +726,15 @@ def test_smoke_audit_requires_frozen_relaxation_md_steps(tmp_path):
         encoding="utf-8",
     )
 
-    report = audit_openmm_smoke_reports(tmp_path)
+    report = audit_openmm_validation_reports(tmp_path)
 
     assert report.status == ValidationStatus.FAIL
-    assert any(check.name == "frozen_protein_relaxation_md_steps" for check in report.checks)
+    assert any(check.name == "conjugate_relaxation_md_steps" for check in report.checks)
 
 
-def test_smoke_audit_accepts_top_level_frozen_relaxation_md_steps(tmp_path):
+def test_openmm_validation_audit_accepts_top_level_conjugate_relaxation_md_steps(tmp_path):
     """Validation should accept top-level MD steps when settings omit them."""
-    (tmp_path / "frozen_protein_relaxation_diagnostics.json").write_text(
+    (tmp_path / "conjugate_relaxation.json").write_text(
         json.dumps(
             {
                 "success": True,
@@ -758,14 +757,14 @@ def test_smoke_audit_accepts_top_level_frozen_relaxation_md_steps(tmp_path):
         encoding="utf-8",
     )
 
-    report = audit_openmm_smoke_reports(tmp_path)
+    report = audit_openmm_validation_reports(tmp_path)
 
     assert report.status == ValidationStatus.PASS
 
 
-def test_smoke_audit_fails_unfixed_frozen_stage_b(tmp_path):
+def test_openmm_validation_audit_fails_unfixed_frozen_stage_b(tmp_path):
     """Validation should fail when Stage B does not keep protein fixed."""
-    diagnostics_path = tmp_path / "frozen_protein_relaxation_diagnostics.json"
+    diagnostics_path = tmp_path / "conjugate_relaxation.json"
     diagnostics_path.write_text(
         json.dumps(
             {
@@ -788,15 +787,15 @@ def test_smoke_audit_fails_unfixed_frozen_stage_b(tmp_path):
         encoding="utf-8",
     )
 
-    report = audit_openmm_smoke_reports(tmp_path)
+    report = audit_openmm_validation_reports(tmp_path)
 
     assert report.status == ValidationStatus.FAIL
-    assert any(check.name == "frozen_protein_relaxation_protein_rmsd" for check in report.checks)
+    assert any(check.name == "conjugate_relaxation_protein_rmsd" for check in report.checks)
 
 
-def test_smoke_audit_requires_frozen_stage_a_success(tmp_path):
+def test_openmm_validation_audit_requires_frozen_stage_a_success(tmp_path):
     """Validation should fail when Stage A minimization is missing."""
-    diagnostics_path = tmp_path / "frozen_protein_relaxation_diagnostics.json"
+    diagnostics_path = tmp_path / "conjugate_relaxation.json"
     diagnostics_path.write_text(
         json.dumps(
             {
@@ -814,10 +813,10 @@ def test_smoke_audit_requires_frozen_stage_a_success(tmp_path):
         encoding="utf-8",
     )
 
-    report = audit_openmm_smoke_reports(tmp_path)
+    report = audit_openmm_validation_reports(tmp_path)
 
     assert report.status == ValidationStatus.FAIL
-    assert any(check.name == "frozen_protein_relaxation_stage_a" for check in report.checks)
+    assert any(check.name == "conjugate_relaxation_stage_a" for check in report.checks)
 
 
 def test_validation_report_status_aggregation(tmp_path):
